chore: move core theme files to providers directory
