+++ src/providers/theme/theme.store.ts
@@ -1,5 +1,5 @@
 import { createStore } from '@stencil/store';
-import { IThemeConfig, ThemeMode, ThemeName } from '../../utils/theme';
+import { IThemeConfig, ThemeMode, ThemeName } from './theme.types';
 
 // eslint-disable-next-line @typescript-eslint/unbound-method
 const { state, onChange, reset, dispose } = createStore<IThemeConfig>({
